gcc- format strings: debug and extra plugins

+++ plugins/protocol_client_loopback_test.c
@@ -151,7 +151,7 @@ callback_client_loopback_test(struct lws *wsi, enum lws_callback_reasons reason,
 
 	case LWS_CALLBACK_CLIENT_RECEIVE:
 		strncpy(buf, in, sizeof(buf) - 1);
-		lwsl_notice("Client connection received %d from server '%s'\n", len, buf);
+		lwsl_notice("Client connection received %ld from server '%s'\n", (long)len, buf);
 
 		/* OK we are done with the client connection */
 		return -1;
